TO KAFKA
